#169 Clean up session commands
