--- src/main/java/scubakay/finalstand/commands/EndSessionCommand.java
@@ -1,27 +1,0 @@
-package scubakay.finalstand.commands;
-
-import com.mojang.brigadier.CommandDispatcher;
-import com.mojang.brigadier.context.CommandContext;
-import net.minecraft.command.CommandRegistryAccess;
-import net.minecraft.server.command.CommandManager;
-import net.minecraft.server.command.ServerCommandSource;
-import scubakay.finalstand.FinalStand;
-import scubakay.finalstand.event.handler.SessionHandler;
-
-public class EndSessionCommand {
-    public static void register(CommandDispatcher<ServerCommandSource> dispatcher, CommandRegistryAccess ignoredCommandRegistryAccess, CommandManager.RegistrationEnvironment ignoredRegistrationEnvironment) {
-        dispatcher.register(
-                CommandManager.literal(FinalStand.COMMAND_ROOT)
-                        .requires(source -> source.hasPermissionLevel(4)) // Must be OP to execute
-                        .then(
-                                CommandManager.literal("end")
-                                        .executes(EndSessionCommand::run)
-                        )
-        );
-    }
-
-    public static int run(CommandContext<ServerCommandSource> context) {
-        SessionHandler.EndSession(context.getSource().getServer());
-        return 1;
-    }
-}
